% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{analyzeBody}
\alias{analyzeBody}
\title{Analyze structure of a function body}
\usage{
analyzeBody(
  callExpr,
  out = c(y.name = FALSE, y.nr = FALSE, y.comp = FALSE, y.vec = FALSE, y.brc = FALSE,
    p.name = FALSE, p.nr = FALSE, p.comp = FALSE, p.vec = FALSE, p.brc = FALSE,
    theta.name = FALSE, theta.nr = FALSE, theta.comp = FALSE, theta.vec = FALSE,
    theta.brc = FALSE, eta.name = FALSE, eta.nr = FALSE, eta.comp = FALSE, eta.vec =
    FALSE, eta.brc = FALSE, eps.name = FALSE, eps.nr = FALSE, eps.comp = FALSE, eps.vec =
    FALSE, eps.brc = FALSE, with = FALSE, return = FALSE, FU = FALSE)
)
}
\arguments{
\item{callExpr}{A call, for example a function body.}

\item{out}{Boolean vector storing the output indicators. Initialized to \code{FALSE} and this
should not be changed.}
}
\value{
Boolean vector of the same format as \code{out}, where \code{TRUE} means the corresponding element
is present in \code{callExpr}, and \code{FALSE} means it is not.
The elements are \code{y.name}, \code{y.nr} and \code{y.comp} indicating whether the vector \code{y}
is indexed by names (strings), numbers or by more complicated expressions (e.g. arithmetic expressions
or functions). The element \code{y.vec} indicates whether \code{y} is referred to as a vector.
The element \code{y.brc} indicates whether elements of \code{y} are referred to with a single bracket ('[]') or '$'.
More than one of these may be true, if multiple reference methods are used in \code{callExpr}.
The elements \code{p.name}, \code{p.nr}, \code{p.comp} and \code{p.vec} do the same for \code{p}, and likewise
for \code{theta}, \code{eta} and \code{eps}.
Elements \code{with}, \code{return} and \code{FU} indicate the presence of the associated functions.
If \code{callExpr} contains illegal components (anything other than a call, pairlist, atomic or name),
then execution is stopped.
}
\description{
Dissects a given function body (technically, any call) to determine if forbidden functions \code{with},
\code{return} or \code{FU_...} occur anywhere, and to determine how the state \code{y} and parameter
\code{p} are indexed.
Functions \code{FU_...} are used internally to represent the state variables and therefore may not be used
for other purposes.
It does these checks by going recursively through the call tree.
This function assumes that \code{y} and \code{p} are the formal arguments representing the state and
parameter vectors.
}
\author{
Martijn van Noort
}
