% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aliasing.r
\name{calcAliasingScoresFromMatrix}
\alias{calcAliasingScoresFromMatrix}
\title{Normalized sensitivities and aliasing scores}
\usage{
calcAliasingScoresFromMatrix(df, vars = NULL)
}
\arguments{
\item{df}{Data frame of variations in tall format (normalized or not), that is, with columns
't', 'i', 'y' and 'dy_d(varsj)', where variables varsj are replaced by names.}

\item{vars}{Vector of names of parameters to be included in aliasing analysis. If \code{NULL} or missing, all parameters in df are used.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A list with two elements, named "normSens" and "aliasingScore".
}\if{html}{\out{</div>}}

The first contains a matrix with as named columns the time "t", output indicator index "i",
and normalized sensitivities of the parameters selected by \code{vars}.
The second contains a matrix with the aliasing scores, one for each pair of parameters
from the selection in \code{vars}.
Each score is a percentage between 0 and 100, where 100 indicates similarity between the parameters
and unidentifiability, and 0 means the parameters are not similar, indicating identifiability.
The matrix is symmetric and the diagonal elements are 100.
In case of errors, the return value is \code{NULL}.
}
\description{
Calculate the normalized sensitivities and aliasing scores for a given variational matrix.
This matrix may be parameter- and/or output-normalized.
Parameter-normalization will not change the outcome (as the sensitivities are normalized anyway), but output-normalization may do so.
}
\details{
Aliasing scores are calculated as follows.
The sensitivity of the model to parameter p_j is s_j(t,i) = dy_i/d_p_j(t), where y_i is the i-th output variable and t is time.
The normalized sensitivity is sn_j(t,i) = s_j(t,i) / max_{t,i}|s_j(t,i)|, and takes values between -1 and 1.
The aliasing score of of parameters p_j and p_k is defined as a(j,k) = 100\% * (1-max_{t,i}(| |sn_j(t,i)| - |sn_k(t,i)| |)),
i.e. it measures the similarity between the profiles of their normalized sensitivities.
In particular a(j,j) = 100\%.
For more details see Augustin et al (2019) and Braakman (2019).
}
\references{
Augustin, Braakman, Paxson - A workflow to detect non-identifiability in parameter estimation
using SimBiology, Github (2019).
Sietse Braakman - A general workflow for parameter estimation to help establish confidence in model
predictions, Rosa Webinar 2019-12-01, slides 39-42.
}
\seealso{
Other calculation functions: 
\code{\link{calcAliasingScoresFromModel}()},
\code{\link{calcFimFromMatrix}()},
\code{\link{calcFimFromModel}()},
\code{\link{calcFirstVariations}()},
\code{\link{calcSensitivityFromMatrix}()},
\code{\link{calcSensitivityFromModel}()},
\code{\link{calcVariationsFim}()},
\code{\link{calcVariations}()},
\code{\link{fimIdent}()}
}
\author{
Martijn van Noort
}
\concept{calculation functions}
