% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{calcFimFromModel}
\alias{calcFimFromModel}
\title{Calculate Fisher Information Matrix}
\usage{
calcFimFromModel(
  model,
  p,
  init,
  output,
  times,
  theta,
  omega,
  sigma,
  vartheta = names(theta),
  varomega = NULL,
  varsigma = NULL,
  symbolic = TRUE,
  chkModel = TRUE,
  ...
)
}
\arguments{
\item{model}{Function(t, y, p) of time, state variables and parameters, specifying the differential equations.
This function should return the numeric vector dy/dt.}

\item{p}{Function(theta, eta) of population and individual parameters specifying the model parameters.
This function should return a named numeric vector.}

\item{init}{Function(p) of parameters specifying the initial state of the model.}

\item{output}{Function(y, p, eps) of state variables, parameters and residual errors specifying the model outputs.
This function should return a numeric vector.}

\item{times}{Numeric vector of times where variations are to be evaluated.}

\item{theta}{Population parameter values, as named numeric vector. May be \code{NULL} if none are needed.}

\item{omega}{Variance-covariance matrix of individual parameters, as numeric matrix.
Should be symmetric and positive definite.
Columns and/or rows should be named. If both, the names should be the same.}

\item{sigma}{Variance-covariance matrix of residual parameters, as numeric matrix.
Should be symmetric and positive definite.
Columns and/or rows should be named. If both, the names should be the same.}

\item{vartheta}{Vector of names of population parameters for which variations are to be calculated.
Should be a subset of \code{names(theta)}, or \code{NULL} for none.
By default equal to \code{names(theta)}.}

\item{varomega}{Matrix (not necessarily named) of booleans of the same dimension as \code{omega}.
The value \code{TRUE} means the derivative of the corresponding element should be included in the FIM,
\code{FALSE} means not.
The matrix should be symmetric.
Instead of a matrix, may also provide a single boolean or \code{NULL}, where \code{TRUE} stands for a
matrix of all \code{TRUE} (also the off-diagonal elements), \code{FALSE} for a matrix of all \code{FALSE},
and \code{NULL} (default) is to set all nonzero elements of \code{omega} to TRUE, and the others to \code{FALSE}.}

\item{varsigma}{Matrix (not necessarily named) of booleans of the same dimension as \code{sigma}.
The value \code{TRUE} means the derivative of the corresponding element should be included in the FIM,
\code{FALSE} means not.
The matrix should be symmetric.
Instead of a matrix, may also provide a single boolean or \code{NULL}, where \code{TRUE} stands for a
matrix of all \code{TRUE} (also the off-diagonal elements), \code{FALSE} for a matrix of all \code{FALSE},
and \code{NULL} (default) is to set all nonzero elements of \code{sigma} to TRUE, and the others to \code{FALSE}.}

\item{symbolic}{\code{TRUE} (default) if derivatives are to be computed symbolically, \code{FALSE} if numerically.
See \code{\link{calcVariations}} for notes on using symbolic derivation.}

\item{chkModel}{\code{TRUE} (default) if it has to be checked whether model components (\code{model}, \code{p}, \code{init}, \code{output})
are formatted correctly for symbolic manipulation, \code{FALSE} if not.}

\item{...}{Named arguments to be passed to \code{\link[deSolve]{lsoda}}.
Can be used for example to pass events or error tolerances.}
}
\value{
The FIM, as a square matrix with element (i,j) equal to -E(d^2 L/d<vi><vj>), where L is the
log likelihood and <vi> is the i-th parameter.
The row and column names of the matrix are those of the variable thetas, omegas and sigmas,
as specified by \code{vartheta}, \code{varomega} and \code{varsigma}.
The FIM is not normalized and the 'thetaNormalized', 'omegaNormalized' and 'sigmaNormalized' attributes
of this matrix are set to named vectors of \code{FALSE}, where the names are the names of the thetas,
and the column names of omegas and sigmas as specified by \code{df}, \code{omega} and \code{sigma},
respectively.
The attributes 'theta', 'omega' and 'sigma' store the values of all parameters as named vectors.
They form a superset of the variable parameters.
The 'type' attribute is set to 'FIM', and the 'symbolic' one to the value of \code{symbolic}.
The function displays an error and returns \code{NULL} if the input is incorrectly formatted or
the variance matrix cannot be inverted.
}
\description{
Calculates the Fisher Information Matrix (FIM) from the given model,
that is, calculates minus the expectation of the second order derivative matrix of the log-likelihood
at the given times with respect to the model parameters.
These parameters are listed in \code{vartheta}, \code{varomega} and \code{varsigma}.
}
\note{
It is recommended to use symbolic derivation (\code{symbolic==TRUE}) rather than numerical derivation,
as it is more accurate, especially for derivatives that are formally zero.
See \code{\link{calcVariations}} for details.
}
\author{
Martijn van Noort
}
