% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{calcFirstVariations}
\alias{calcFirstVariations}
\title{Calculate first order variations}
\usage{
calcFirstVariations(
  model,
  parms,
  init,
  outputPred,
  times,
  varp = names(parms),
  symbolic = TRUE,
  chkModel = TRUE,
  ...
)
}
\arguments{
\item{model}{Function(t, y, p) of time, state variables and parameters, specifying the differential equations.
This function should return the numeric vector dy/dt.}

\item{parms}{Parameter values, as named numeric vector. May be \code{NULL} if none are needed
(but then no derivatives will be calculated).}

\item{init}{Function(p) of parameters specifying the initial state (at time 0) of the model.}

\item{outputPred}{Function(y, p) of state variables and parameters specifying the model prediction, i.e., the
output without residual.
This function should return a numeric vector.}

\item{times}{Numeric vector of times where variations are to be evaluated.
Should all be >= 0, but do not need to include the dose time 0.}

\item{varp}{Vector of names of parameters for which variations are to be calculated.
Should be a subset of \code{names(parms)}, or \code{NULL} for none.
By default equal to \code{names(parms)}.}

\item{symbolic}{\code{TRUE} (default) if derivatives are to be computed symbolically, \code{FALSE} if numerically.
See \code{\link{calcVariations}} for details.}

\item{chkModel}{\code{TRUE} (default) if it has to be checked whether model components (\code{model}, \code{parms}, \code{init}, \code{outputPred})
are formatted correctly for symbolic manipulation, \code{FALSE} if not.}

\item{...}{Named arguments to be passed to \code{\link[deSolve]{lsoda}}.
Can be used for example to pass events or error tolerances.}
}
\value{
A data frame with columns 't' for time, 'i' for the index of the output element, 'y' for the outputs,
and 'dy/d<v1>' for the first derivatives.
In the column names, variable v1 is replaced by the names in \code{varp}.
The data frame has an attribute 'theta' listing the variables in \code{parms} as a named vector,
and 'eta' and 'eps' that are empty.
The 'type' attribute is set to 'VariationalMatrix', 'secOrd' to \code{FALSE}, and 'symbolic' to the value of
\code{symbolic}.
The variational matrix is not normalized, as reflected in the attributes 'parameterNormalized' and 'outputNormalized',
that record input and output normalization, respectively.
The attribute 'parameterNormalized' is set to a named boolean vector of \code{FALSE}, with the \code{parms}
names as names.
The attribute 'outputNormalized' is set to \code{FALSE}.
The function displays an error and returns \code{NULL} if \code{varp} contains elements not in \code{names(p)},
or if \code{times} includes a negative value.
}
\description{
Calculates the first variational matrix of the given model, that is, the first order derivatives of
the model outputs with respect to the model parameters, at given times.
This is tailored for use with the sensitivity and aliasing methods.
Therefore model parameters are assumed to be given directly, rather than as a combination of population and
individual parameters, and residual error parameters are not included.
}
\note{
It is recommended to use symbolic derivation (\code{symbolic==TRUE}) rather than numerical derivation,
as it is more accurate, especially for derivatives that are formally zero.
See \code{\link{calcVariations}} for details.
}
\author{
Martijn van Noort
}
