% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{calcVariationsFim}
\alias{calcVariationsFim}
\title{Calculate first and second order variations for FIM}
\usage{
calcVariationsFim(
  model,
  p,
  init,
  output,
  times,
  theta,
  nmeta,
  nmeps,
  vartheta = names(theta),
  vareta = nmeta,
  vareps = nmeps,
  symbolic = TRUE,
  chkModel = TRUE,
  ...
)
}
\arguments{
\item{model}{Function(t, y, p) of time, state variables and parameters, specifying the differential equations.
This function should return the numeric vector dy/dt.}

\item{p}{Function(theta, eta) of population and individual parameters specifying the model parameters.
This function should return a named numeric vector.}

\item{init}{Function(p) of parameters specifying the initial state (at time 0) of the model.}

\item{output}{Function(y, p, eps) of state variables, parameters and residual errors specifying the model outputs.
This function should return a numeric vector.}

\item{times}{Numeric vector of times where variations are to be evaluated.
Should all be >= 0, but do not need to include the dose time 0.}

\item{theta}{Population parameter values, as named numeric vector. May be \code{NULL} if none are needed.}

\item{nmeta}{Names of individual parameters, as string vector. May be \code{NULL} if none are needed.}

\item{nmeps}{Names of residual parameter values, as string vector. May be \code{NULL} if none are needed.}

\item{vartheta}{Vector of names of population parameters for which variations are to be calculated.
Should be a subset of \code{names(theta)}, or \code{NULL} for none.
By default equal to \code{names(theta)}.}

\item{vareta}{Vector of names of individual parameters for which variations are to be calculated.
Should be a subset of \code{nmeta}, or \code{NULL} for none.
By default equal to \code{nmeta}.}

\item{vareps}{Vector of names of residual parameters for which variations are to be calculated.
Should be a subset of \code{nmeps}, or \code{NULL} for none.
By default equal to \code{nmeps}.}

\item{symbolic}{\code{TRUE} (default) if derivatives are to be computed symbolically, \code{FALSE} if numerically.
See \code{\link{calcVariations}} for details.}

\item{chkModel}{\code{TRUE} (default) if it has to be checked whether model components (\code{model}, \code{p}, \code{init}, \code{output})
are formatted correctly for symbolic manipulation, \code{FALSE} if not.}

\item{...}{Named arguments to be passed to \code{\link[deSolve]{lsoda}}.
Can be used for example to pass events or error tolerances.}
}
\value{
A data frame with columns 't' for time, 'i' for the index of the output element, 'y' for the outputs,
'dy/d<v1>' for the first derivatives and 'd2y/d<v1>_<v2>' for the second derivatives.
In the column names, variables v1 and v2 are replaced by the names in \code{vartheta}, \code{vareta} and \code{vareps}.
The data frame has an attribute 'theta' listing the values of \code{theta} as a named vector, and attributes
'eta' and 'eps' containing named vectors with 0 values for all parameters in \code{nmeta} and \code{nmeps}.
The 'type' attribute is set to 'VariationalMatrix', 'secOrd' to \code{TRUE} and 'symbolic' to the value of
\code{symbolic}.
The variational matrix is not normalized, as reflected in the attributes 'parameterNormalized' and 'outputNormalized',
that record input and output normalization, respectively.
The attribute 'parameterNormalized' is set to a named boolean vector of \code{FALSE}, with the theta, eta and eps
parameter names as names.
The attribute 'outputNormalized' is set to \code{FALSE}.
The function displays an error and returns \code{NULL} if \code{vartheta} contains elements not in \code{names(theta)},
and likewise for \code{nmeta} and \code{nmeps}, or if \code{times} includes a negative value.
}
\description{
Calculates the variational matrix of the given model, that is, the first and second order derivatives of
the model outputs with respect to the model parameters, at given times.
This is tailored for use with the FIM methods.
Therefore derivatives of individual and residual error parameters are evaluated at 0, and so only their
names need to be provided as input.
}
\note{
The inputs \code{vareta} and \code{vareps} should contain all random parameters with nonzero variance, even if the
FIM calculation considers the variance and covariances of such a parameter as fixed (that is, derivatives with respect to
these variance and covariances do not appear in the FIM).
That is, random parameters may be dropped from \code{vareta} and \code{vareps} only if they have zero variance.
Notice that any variance and covariance parameters appearing in the FIM necessarily correspond to random parameters with
nonzero variance, so inclusion of all random parameters with nonzero variance is not only necessary but also sufficient.
That is, \code{vareta} and \code{vareps} can be set equal to the sets of all individual and residual parameters with
nonzero variance, respectively.

It is recommended to use symbolic derivation (\code{symbolic==TRUE}) rather than numerical derivation,
as it is more accurate, especially for derivatives that are formally zero.
See \code{\link{calcVariations}} for details.
}
\seealso{
Other calculation functions: 
\code{\link{calcAliasingScoresFromMatrix}()},
\code{\link{calcAliasingScoresFromModel}()},
\code{\link{calcFimFromMatrix}()},
\code{\link{calcFimFromModel}()},
\code{\link{calcFirstVariations}()},
\code{\link{calcSensitivityFromMatrix}()},
\code{\link{calcSensitivityFromModel}()},
\code{\link{calcVariations}()},
\code{\link{fimIdent}()}
}
\author{
Martijn van Noort
}
\concept{calculation functions}
