% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{calcVariationsSymb}
\alias{calcVariationsSymb}
\title{Calculate variations symbolically}
\usage{
calcVariationsSymb(
  model,
  p,
  init,
  output,
  times,
  theta,
  eta,
  eps,
  vartheta,
  vareta,
  vareps,
  secOrd,
  chkModel,
  showWarn,
  ...
)
}
\arguments{
\item{model}{Function(t, y, p) of time, state variables and parameters, specifying the differential equations.
This function should return the numeric vector dy/dt.}

\item{p}{Function(theta, eta) of population and individual parameters specifying the model parameters.
This function should return a named numeric vector.}

\item{init}{Function(p) of parameters specifying the initial state (at time 0) of the model.}

\item{output}{Function(y, p, eps) of state variables, parameters and residual errors specifying the model outputs.
This function should return a numeric vector.}

\item{times}{Numeric vector of times where variations are to be evaluated.
Should all be >= 0, but do not need to include the dose time 0.}

\item{theta}{Population parameter values, as named numeric vector. May be \code{NULL} if none are needed.}

\item{eta}{Individual parameter values, as named numeric vector. May be \code{NULL} if none are needed.}

\item{eps}{Residual parameter values, as named numeric vector. May be \code{NULL} if none are needed.}

\item{vartheta}{Vector of names of population parameters for which variations are to be calculated.
Should be a subset of \code{names(theta)}, or \code{NULL} for none.}

\item{vareta}{Vector of names of individual parameters for which variations are to be calculated.
Should be a subset of \code{names(eta)}, or \code{NULL} for none.}

\item{vareps}{Vector of names of residual parameters for which variations are to be calculated.
Should be a subset of \code{names(eps)}, or \code{NULL} for none.}

\item{secOrd}{\code{TRUE} if second order derivatives have to be computed, \code{FALSE} if not.}

\item{chkModel}{\code{TRUE} if it has to be checked whether model components (\code{model}, \code{p}, \code{init}, \code{output})
are formatted correctly for symbolic manipulation, \code{FALSE} if not.}

\item{showWarn}{\code{TRUE} if warnings about missing random parameters are to be shown, \code{FALSE} if not.
The latter is of use in case only the structural model is of interest.}

\item{...}{Named arguments to be passed to \code{\link[deSolve]{lsoda}}.
Can be used for example to pass events or error tolerances.}
}
\value{
A data frame with columns 't' for time, 'i' for the index of the output element, 'y' for the outputs,
'dy_d<v1>' for the first derivatives and (if \code{secOrd == TRUE}) 'd2y_d<v1>_<v2>' for the second derivatives.
In the column names, variables v1 and v2 are replaced by the names in \code{vartheta}, \code{vareta} and \code{vareps}.
The function displays an error and returns \code{NULL} if \code{vartheta} contains elements not in \code{names(theta)},
and likewise for \code{eta} and \code{eps}.
The same happens if the model components are not formatted correctly for symbolic manipulation and \code{chkModel = TRUE},
or if integration of the model fails.
}
\description{
Calculates the variational matrix of the given model, that is, the first and optionally second order derivatives of
the model outputs with respect to the model parameters, at given times.
}
\author{
Martijn van Noort
}
