% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{calcVarSymb}
\alias{calcVarSymb}
\title{Calculate output in case of symbolic derivatives}
\usage{
calcVarSymb(doutput, dpout, vartheta, vareta, vareps, secOrd)
}
\arguments{
\item{doutput}{data frame containing values of derivatives of order 0, 1 and optionally 2 of the output
vector, each row in the format of the output of the function \code{doutput}, see \code{\link{createDerivsSymb}},
and each row corresponding to a time point.
The column names should reflect the component and derivative, as in the output of the function \code{doutput}.}

\item{dpout}{named numeric list containing values of derivatives of order 0, 1 and optionally 2 of the parameter
vector, in the format of the output of the function \code{dp}, see \code{\link{createDerivsSymb}}.}

\item{vartheta}{Vector of names of population parameters for which variations are to be calculated.
Should be a subset of \code{names(theta)}, or \code{NULL} for none.}

\item{vareta}{Vector of names of individual parameters for which variations are to be calculated.
Should be a subset of \code{names(eta)}, or \code{NULL} for none.}

\item{vareps}{Vector of names of residual parameters for which variations are to be calculated.
Should be a subset of \code{names(eps)}, or \code{NULL} for none.}

\item{secOrd}{\code{TRUE} if second order derivatives have to be computed, \code{FALSE} if not.}
}
\value{
A vector with derivatives of order 0, 1, 2 of the state variables with respect to base parameters
(theta, eta and eps). The elements are named <nam>.<i>, where i is the sequence number of the component of y,
and nam is of the form 'y', 'dy_d<v1>' or 'd2y_d<v1>_<v2>'.
}
\description{
Calculates the variations in case of symbolic derivatives, for a given output and parameter vector, both in variational form,
that is, including derivatives.
For the output vector these are derivatives with respect to the derived parameters (\code{p}).
For the parameter vector these are derivatives with respect to the base parameters (\code{theta}, \code{eta}, \code{eps}),
denoted here as \code{v}.
This function combines these to obtain derivatives of the output with respect to the base parameters.
This is an applicaton of the chain rule for derivatives: dy_i/dv = sum_j(dy_i/dp_j dp_j/dv) and
d^2y_i/(dv dw) = sum_{j,k}(d^2y_i/(dp_j dp_k) dp_j/dv dp_k/dv) + sum_j(dy_i/dp_j d^2p_j/(dv dw)).
}
