% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{checkModelDef}
\alias{checkModelDef}
\title{Check formatting of model components for symbolic manipulation}
\usage{
checkModelDef(model, p, init, output, theta, eta, eps, showWarn)
}
\arguments{
\item{model}{Function(t, y, p) of time, state variables and parameters, specifying the differential equations.
This function should return the numeric vector dy/dt.}

\item{p}{Function(theta, eta) of population and individual parameters specifying the model parameters.
This function should return a named numeric vector.}

\item{init}{Function(p) of parameters specifying the initial state (at time 0) of the model.}

\item{output}{Function(y, p, eps) of state variables, parameters and residual errors specifying the model outputs.
This function should return a numeric vector.}

\item{theta}{Population parameter values, as named numeric vector. May be \code{NULL} if none are needed.}

\item{eta}{Individual parameter values, as named numeric vector. May be \code{NULL} if none are needed.}

\item{eps}{Residual parameter values, as named numeric vector. May be \code{NULL} if none are needed.}

\item{showWarn}{\code{TRUE} if warnings about missing random parameters are to be shown, \code{FALSE} if not.
The latter is of use in case only the structural model is of interest.}
}
\value{
\code{TRUE} if the model components pass all tests, \code{FALSE} if not.
In the latter case, error messages will be displayed.
All encountered errors are shown, but if the formals are incorrect, the other checks are not performed.
In the first case, warning messages may be displayed.
}
\description{
Checks whether model components (\code{model}, \code{p}, \code{init}, \code{output})
are formatted correctly for symbolic manipulation.
The following checks are performed:
\itemize{
\item Each component has the correct formals.
\item The components do not contain "with" or "return" statements, or functions starting with "FU_".
The latter are used to represent the state variables in internal calculations and therefore may not be used for other purposes.
\item The components do not refer to y, p, theta, eta or eps as vector or indexed via "\link{.}".
\item The components refer to elements of y, p, theta, eta or eps in the same way, i.e., either by name or number.
\item The vectors \code{theta}, \code{eta} and \code{eps} are consistent with this (i.e. whether they name their components or not).
\item The outputs of the functions \code{init} and \code{p} are consistent with this (i.e. whether they name their outputs or not).
}
}
\note{
\code{theta}, \code{eta} and \code{eps} are only used to generate some output values of the functions
\code{p} and \code{init}, so that their naming can be checked, and to check whether they have names or not.
Their specific values usually do not matter (but see next note).

These checks are meant to catch the most common mistakes but are not completely failsafe. For example,
to check consistency in naming they determine if \code{p} and \code{init} provide named or unnamed return values for
the given \code{theta} and \code{eta}. This may fail to detect errors in case the naming depends on the parameter values.
Another example is that the validity or consistency of names or indices is not checked
(e.g., one model component may use different names than another, or indices can be negative).
}
\author{
Martijn van Noort
}
