% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{checkPosDef}
\alias{checkPosDef}
\title{Check whether a matrix is positive definite}
\usage{
checkPosDef(mat, tol = 1e-08)
}
\arguments{
\item{mat}{The matrix to check.}

\item{tol}{The check for positive definiteness tests whether all eigenvalues have a real part larger than \code{tol}.}
}
\value{
A string with an error message if not positive definite or not square,
and \code{NULL} if positive definite. The empty matrix is by definition positive definite .
}
\description{
Checks whether the matrix \code{mat} is positive definite.
}
\author{
Martijn van Noort
}
