% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{checkSetCovMat}
\alias{checkSetCovMat}
\title{Check variance-covariance matrices and return them in a standard format}
\usage{
checkSetCovMat(omega, sigma, varomega, varsigma, caller)
}
\arguments{
\item{omega}{The variance-covariance matrix of the individual effects, provided as a numeric matrix.}

\item{sigma}{The variance-covariance matrix of the residual effects, provided as a numeric matrix.}

\item{varomega}{Indicators which elements of \code{omega} should be considered as variable, provided
as a boolean matrix or as a single boolean or \code{NULL}.
In case of a single boolean, it is assumed that this single value applies to every matrix element
(also the off-diagonal ones).
The value \code{NULL} means all nonzero elements of \code{omega} are TRUE, and the others are \code{FALSE}.}

\item{varsigma}{Indicators which elements of \code{sigma} should be considered as variable, provided
as a boolean matrix or as a single boolean or \code{NULL}.
In case of a single boolean, it is assumed that this single value applies to every matrix element
(also the off-diagonal ones).
The value \code{NULL} means all nonzero elements of \code{sigma} are TRUE, and the others are \code{FALSE}.}

\item{caller}{String containing the signature of the calling function, for use in error messages.
A typical value would be 'FisherInfo::calcFimFromMatrix'.}
}
\value{
\code{NULL} if there were errors. In this case an error message will be printed.
Otherwise a list of four elements.
The first two are the variance-covariance matrices, i.e. \code{omega} and \code{sigma}.
The second pair are the indicator matrices \code{varomega} and \code{varsigma}.
All four are adapted from input as described above.
}
\description{
Performs several checks on the variance-covariance matrix inputs of the methods that calculate the FIM,
and return them in a standard format.
Can also be used to standardize the input to \code{\link{normalizeFim}}.
For the inputs \code{omega} and \code{sigma} it is checked that they are matrices
that are square, symmetric, positive definite and with matching row and column names.
Those names are set equal if necessary.
For the inputs \code{varomega} and \code{varsigma} it is checked that they are symmetric matrices,
or single values from which symmetric matrices are constructed.
Their row and column names are set to match those of \code{omega} and \code{sigma}.
}
\details{
Superfluous rows and columns are removed from these matrices.
That is, any row (and corresponding column) with only zero variances and covariances in \code{omega} and
only \code{FALSE} in \code{varomega} will be removed from both matrices, as it would not affect the outcome.
Likewise for \code{sigma} and \code{varsigma}.
The resulting matrices \code{omega} and \code{sigma} should not both have dimension zero.
}
\author{
Martijn van Noort
}
