% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{checkSetDerivCols}
\alias{checkSetDerivCols}
\title{Check that the matrix \code{df} contains all required derivatives and return their column names}
\usage{
checkSetDerivCols(df, omega, sigma, vartheta, caller)
}
\arguments{
\item{df}{Data frame of first and second order variations as generated by \code{\link{calcVariations}}
or \code{\link{calcVariationsFim}}, that is, with columns 't', 'i', 'y', 'dy_d<v1>' and
'd2y_d<v1>_<v2>', where variables v1 and v2 are replaced by names.}

\item{omega}{The variance-covariance matrix of the individual effects, provided as a numeric matrix.}

\item{sigma}{The variance-covariance matrix of the residual effects, provided as a numeric matrix.}

\item{vartheta}{Vector of names of population parameters for which derivatives are to be calculated.
Should be a subset of the names appearing in the variational matrix, or \code{NULL} for none.}

\item{caller}{String containing the signature of the calling function, for use in error messages.
A typical value would be 'FisherInfo::calcFimFromMatrix'.}
}
\value{
\code{NULL} if there were errors. In this case an error message will be printed.
Otherwise a list of five elements, named "theta", "eta", "eps", "thetaEta" and "thetaEps" containing
the column names in \code{df} of the derivatives of the indicated type.
}
\description{
The FIM calculation requires all first derivatives "dy_d<p>" and all second derivatives "d2y_dp_q" or "d2y_dq_p",
where p can be any element of \code{vartheta} or any rowname of \code{omega} or \code{sigma},
and q can be any rowname of \code{omega} or \code{sigma}.
}
\details{
This function checks that the matrix \code{df} contains all these derivatives, and returns their names in \code{df}.
It also checks that the derivative names are returned in canonical order, i.e., corresponding to the order in
omega and sigma.
}
\author{
Martijn van Noort
}
