% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{checkSetSquareAndNames}
\alias{checkSetSquareAndNames}
\title{Check whether the input is a square matrix and set its row or column names}
\usage{
checkSetSquareAndNames(mat)
}
\arguments{
\item{mat}{The object to check}
}
\value{
a list of two elements.
The first is the matrix \code{mat} with row or column names adapted as described above
(if there were no errors).
The second is an error message (as a string), or \code{NULL} if there were no errors.
}
\description{
Checks whether the input \code{mat} is a square matrix (or \code{NULL}), and whether it has row and/or column names.
If it has both row and column names, checks whether they are the same.
If it has one of the two, sets the other equal to it.
Returns the adapted matrix and a list of errors.
}
\author{
Martijn van Noort
}
