% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{checkSymm}
\alias{checkSymm}
\title{Check whether a matrix is symmetric}
\usage{
checkSymm(mat)
}
\arguments{
\item{mat}{The matrix to check.}
}
\value{
A string with an error message if not symmetric,
and \code{NULL} if symmetric.
}
\description{
Checks whether the matrix \code{mat} is symmetric.
}
\author{
Martijn van Noort
}
