% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{checkVectorUse}
\alias{checkVectorUse}
\title{check vector use in a function body}
\usage{
checkVectorUse(callExpr, atTailHeadBlock = TRUE)
}
\arguments{
\item{callExpr}{A call, for example a function body.}

\item{atTailHeadBlock}{Boolean vector storing whether this call is at the tail of the function, i.e.,
whether it contains the return statement.
Initialized to \code{TRUE} and this should not be changed.}
}
\value{
\code{TRUE} if \code{c} or \code{`:`} occur anywhere outside the return statement, and \code{FALSE}
if they only occur in the return value (or not at all).
If \code{callExpr} contains illegal components (anything other than a call, pairlist, atomic or name),
then the return value is \code{NA}. This should never happen.
}
\description{
Dissects a given function body (technically, any call) to determine if forbidden vectorizing functions \code{c}
or \code{`:`} occur anywhere outside the return value.
It does so by going recursively through the call tree.
}
\author{
Martijn van Noort
}
