% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{computeNormalizers}
\alias{computeNormalizers}
\title{Normalizing factors for variance matrix}
\usage{
computeNormalizers(mat, relvec, existNorm)
}
\arguments{
\item{mat}{Variance-covariance matrix, as numeric matrix.
Should be symmetric and positive definite.
Columns and rows should be named, and the names should be the same.}

\item{relvec}{Boolean vector of the same dimension as \code{mat},
indicating which elements need to be normalized (\code{TRUE}) or not (\code{FALSE}).
Elements need not be named.}

\item{existNorm}{Named boolean vector specifying the existing normalization.
This should have the same dimension as \code{mat}, and the names should equal the
column names of \code{mat}, in the same order.}
}
\value{
A data frame with variables 'row', 'col' and 'nam' specifying the matrix element,
and 'value' containing the normalizing factor.
This data frame may be empty if \code{mat} is empty.
The function displays an error and returns \code{NULL} if \code{existNorm} is not in the specified format.
}
\description{
Computes normalizing factors for a given variance-covariance matrix.
The factor takes into account prior normalization.
This function is for internal use and performs no validation.
}
\author{
Martijn van Noort
}
