% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{copyEnv}
\alias{copyEnv}
\title{Copy an environment}
\usage{
copyEnv(env)
}
\arguments{
\item{env}{Environment to copy.}
}
\value{
A copy of this environment.
}
\description{
Creates a shallow copy of the given environment.
Changes to this copy will not affect the original environment.
In this sense, \code{newEnv <- copyEnv(env)} is different from \code{newEnv <- env}.
Designed to create a copy of the rules environment of \code{\link[Deriv]{Deriv}}.
}
\author{
Martijn van Noort
}
