% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{createDerivGreps}
\alias{createDerivGreps}
\title{Create grep strings for derivatives.}
\usage{
createDerivGreps(
  derivCodes,
  compNames,
  varName,
  derivSep = "\\\\.",
  typeSep = "_",
  outnameSep = "_"
)
}
\arguments{
\item{derivCodes}{Vector of strings containing the derivative codes.
Elements have the form \code{0} for the 0-th derivative, \code{1<derivSep><deriv>}
for the 1-st derivatives, and \code{2<derivSep><deriv><derivSep><deriv>} for the second
derivatives. Here \code{<deriv>} codes a derivative as \code{<type><typeSep><variable>},
where \code{<type>} is a derivative type, such as "theta" or "eta", and \code{<variable>}
is a variable (from the list of variables of the given type), such as "TVCL".}

\item{compNames}{Vector of strings specifying the names of the output components of the 0-th derivative,
or \code{NULL} if they have no names.}

\item{varName}{String specifying the variable name to be used in the output.}

\item{derivSep}{String specifying the character(s) separating the various parts of an element of \code{derivCodes}.}

\item{typeSep}{String specifying the character(s) separating the type and variable in elements of \code{derivCodes}.}

\item{outnameSep}{String specifying the character(s) separating the various parts of the derivative specification
in the names of the output.}
}
\value{
The grep strings, as a named vector of strings of the same length as derivCodes.
The names of its elements are \code{<varName><outnameSep><deriv1><outnameSep><deriv2>}, where the derivatives are
included as needed.
}
\description{
Creates grep strings for selecting the given derivatives from the output of \code{\link[Deriv]{Deriv}}.
To this end, the codes provided in \code{derivCodes} are translated to match the names of elements of this output.
The grep strings take into account that these names can be different, depending on simplifications of the output.
That is, output vectors equal to 0 are sometimes reduced to single elements, and corresponding parts of their
name are suppressed.
Also, derivatives may be wholly or partially suppressed.
}
\author{
Martijn van Noort
}
