% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{createDerivsSymb}
\alias{createDerivsSymb}
\title{Generate derivatives of model components symbolically}
\usage{
createDerivsSymb(
  model,
  p,
  init,
  output,
  theta,
  eta,
  eps,
  vartheta,
  vareta,
  vareps,
  secOrd
)
}
\arguments{
\item{model}{Function(t, y, p) of time, state variables and parameters, specifying the differential equations.
This function should return the numeric vector dy/dt.}

\item{p}{Function(theta, eta) of population and individual parameters specifying the model parameters.
This function should return a named numeric vector.}

\item{init}{Function(p) of parameters specifying the initial state (at time 0) of the model.}

\item{output}{Function(y, p, eps) of state variables, parameters and residual errors specifying the model outputs.
This function should return a numeric vector.}

\item{theta}{Population parameter values, as named numeric vector. May be \code{NULL} if none are needed.}

\item{eta}{Individual parameter values, as named numeric vector. May be \code{NULL} if none are needed.}

\item{eps}{Residual parameter values, as named numeric vector. May be \code{NULL} if none are needed.}

\item{vartheta}{Vector of names of population parameters for which variations are to be calculated.
Should be a subset of \code{names(theta)}, or \code{NULL} for none.}

\item{vareta}{Vector of names of individual parameters for which variations are to be calculated.
Should be a subset of \code{names(eta)}, or \code{NULL} for none.}

\item{vareps}{Vector of names of residual parameters for which variations are to be calculated.
Should be a subset of \code{names(eps)}, or \code{NULL} for none.}

\item{secOrd}{\code{TRUE} if second order derivatives have to be computed, \code{FALSE} if not.}
}
\value{
A list of four functions, containing the variational versions of \code{model}, \code{p}, \code{init} and
\code{output}. These functions take same arguments as their non-variational counterparts, but with a different
meaning for the argument \code{y}, namely this contains the state variables and their derivatives wrt parameters.
The output of \code{p} is a named list of named numeric vectors, with one element for each derivative.
The names of the elements have the format \code{p(_<deriv>(_<deriv>))}, where \code{<deriv>} codes the derivative and
is optional, as indicated by the brackets.
Each element has the same length as the output of the non-variational version of \code{p}, and its components have
the same names.
The output of \code{init} and \code{model} is a named numeric vector of values of the state variables and their
derivatives.
The names have the format \code{y(_<deriv>(_<deriv>)).<state>}, where \code{<deriv>} codes the derivative and
is optional, and \code{<state>} codes the state.
The output of \code{output} is a named numeric vector of values of the output variables and their derivatives.
The names have the format \code{y(_<deriv>(_<deriv>)).<state>}, where \code{<deriv>} codes the derivative and
is optional, and \code{<state>} codes the state or (if the state has no name) its index.
}
\description{
Calculates first and optionally second derivatives of the model, initialization, output and parameter functions
with respect to relevant parameters, in a symbolic way.
}
\author{
Martijn van Noort
}
