% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{createLowerTriSecondDerivCodes}
\alias{createLowerTriSecondDerivCodes}
\title{Create second derivative codes}
\usage{
createLowerTriSecondDerivCodes(firstDerivCodes, sep = ".")
}
\arguments{
\item{firstDerivCodes}{Vector of strings containing codes of first derivatives.}

\item{sep}{String defining the separator character(s) between the two derivatives in the output.
Default is ".".}
}
\value{
Vector of strings containing second derivative codes, of the form \code{<deriv1><sep><deriv2>}, where
\code{<deriv1>} and \code{<deriv2>} are elements of \code{firstDerivCodes}.
Returns \code{NULL} if there are none. This happens when \code{firstDerivCodes} is empty or NULL.
}
\description{
Creates second derivative codes from first derivative ones in lower triangular form, that is, codes are generated
for derivatives (i,j) where i <= j.
}
\author{
Martijn van Noort
}
