% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{createOmegaSigmaCases}
\alias{createOmegaSigmaCases}
\title{Create data frame indicating which derivatives of random effects should be taken}
\usage{
createOmegaSigmaCases(varmat, isOmega)
}
\arguments{
\item{varmat}{A named square boolean matrix indicating which elements of a variance-covariance matrix
should be considered as variable. Row names should equal column names.}

\item{isOmega}{\code{TRUE} if the matrix \code{varmat} corresponds to individual effects,
\code{FALSE} if residual effects.}
}
\value{
Data frame with cases.
It contains numeric columns "row" and "col" indicating the row and column index in \code{varmat} of the case,
a string column "nam" providing a description of the case, and a boolean column "omega" equal to \code{isOmega}.
The column index is never higher than the row index.
If row = col, then "nam" is set to the parameter name (i.e., the row name of \code{varmat}), and otherwise it
is set to "(<par1>,<par2>)", where <par1> and <par2> are the parameter names of row and column, respectively.
The data frame may be empty if there were no cases to list.
}
\description{
Creates a data frame with on each row a case description of a derivative of a random effect,
given a matrix indicating the variability of elements of a variance-covariance matrix.
}
\author{
Martijn van Noort
}
