% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{extractDf}
\alias{extractDf}
\title{Extract columns from a data frame}
\usage{
extractDf(df, cols)
}
\arguments{
\item{df}{data frame}

\item{cols}{vector of strings specifying columns to be extracted}
}
\value{
A data frame containing the required columns
}
\description{
Extracts the specified columns from a data frame.
The difference with \code{[]} is that missing columns are replaced by zeroes.
}
