% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{getAllParsFim}
\alias{getAllParsFim}
\title{Get all parameter values of a FIM}
\usage{
getAllParsFim(fim)
}
\arguments{
\item{fim}{The FIM.}
}
\value{
A named list of three named numeric vectors containing the theta's, omega's and sigma's used to
create the FIM.
The list elements are named "theta", "omega" and "sigma", respectively.
The vector elements have the names of the parameters, where off-diagonal elements of omega and sigma are
named "(<par1>,<par2>)".
The function displays an error and returns \code{NULL} if \code{fim} is not a valid FIM.
}
\description{
Gets the values of the parameters used to create a given FIM.
}
\seealso{
Other retrieval functions: 
\code{\link{[.keepattr}()},
\code{\link{getAllParsVariations}()},
\code{\link{getParVecFim}()},
\code{\link{getParVecVariations}()},
\code{\link{getParsFim}()},
\code{\link{getParsVariations}()},
\code{\link{subFim}()}
}
\author{
Martijn van Noort
}
\concept{retrieval functions}
