% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{getAllParsVariations}
\alias{getAllParsVariations}
\title{Get all parameter values of a variational matrix}
\usage{
getAllParsVariations(df)
}
\arguments{
\item{df}{A data frame containing a variational matrix, optionally with second order derivatives.}
}
\value{
A named list of three named numeric vectors containing the theta's, eta's and eps's used to
create the variational matrix.
The list elements are named "theta", "eta" and "eps", respectively.
The vector elements have the names of the parameters.
The function displays an error and returns \code{NULL} if \code{df} is not a valid variational matrix.
}
\description{
Gets all parameter values that were used to create a given variational matrix.
}
\author{
Martijn van Noort
}
