% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{getParsVariations}
\alias{getParsVariations}
\title{Get column parameter values of a variational matrix}
\usage{
getParsVariations(df)
}
\arguments{
\item{df}{A data frame containing a variational matrix, optionally with second order derivatives.}
}
\value{
A named list of three named numeric vectors containing the theta's, eta's and eps's used to
create the variational matrix and present as columns of the matrix, in the order of appearance in the matrix.
The list elements are named "theta", "eta" and "eps", respectively.
The vector elements have the names of the parameters.
The function displays an error and returns \code{NULL} if \code{df} is not a valid variational matrix.
}
\description{
Gets the parameter values of the columns of a variational matrix that were used to create a given variational matrix,
as a list of population, individual and residual parameters.
This is the subset of \code{getAllParsVariations} corresponding to the columns present in the variational matrix.
The output is the same as that of \code{getParVecVariations} except it is formatted as a list of three vectors.
}
\author{
Martijn van Noort
}
