% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{isNormalizedFim}
\alias{isNormalizedFim}
\title{Checks whether a FIM is normalized}
\usage{
isNormalizedFim(fim)
}
\arguments{
\item{fim}{A matrix with row and column names containing the FIM,
as produced by \code{\link{calcFimFromMatrix}} or \code{\link{calcFimFromModel}}.}
}
\value{
a named boolean vector, stating for each parameter of the FIM whether it is
normalized (\code{TRUE}) or not (\code{FALSE}).
All theta, omega and sigma parameters are listed, not just the ones present in the FIM.
Returns \code{NULL} and prints an error if there was no normalization indicator.
}
\description{
Returns whether the given FIM is normalized, as recorded in its normalization indicators.
}
\seealso{
Other checkers: 
\code{\link{isNormalizedVariations}()}
}
\author{
Martijn van Noort
}
\concept{checkers}
