% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{isNormalizedVariations}
\alias{isNormalizedVariations}
\title{Check whether a variational matrix is normalized}
\usage{
isNormalizedVariations(df)
}
\arguments{
\item{df}{A data frame containing a variational matrix, optionally with second order derivatives.}
}
\value{
A named boolean list stating the parameter- or output-normalization of the matrix.
Function displays an error and returns \code{NULL} if the input is not in the specified format.
}
\description{
Checks whether a variational matrix is normalized.
The check involves checking the validity of the input and the value of its attributes \code{parameterNormalized}
and \code{outputNormalized}.
}
\seealso{
Other checkers: 
\code{\link{isNormalizedFim}()}
}
\author{
Martijn van Noort
}
\concept{checkers}
