% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{isValidFim}
\alias{isValidFim}
\title{Check validity of a FIM}
\usage{
isValidFim(df, nmdf = "input 'df'")
}
\arguments{
\item{df}{Any object.}

\item{nmdf}{Name (string) by which this object should be referred in any error messages. By default "input 'df'".}
}
\value{
\code{NULL} if \code{df} is a valid FIM,
and an error message or list of error messages if not.
}
\description{
Checks whether the provided input is a valid FIM,
that is, whether it is in the format as produced by \code{\link{calcFimFromMatrix}} or \code{\link{calcFimFromModel}}.
In particular, it is checked whether the input is a matrix of numeric values, with attributes 'type', 'symbolic',
'omega', 'sigma', 'thetaNormalized', 'omegaNormalized' and 'sigmaNormalized', where 'type' is set to 'FIM' and
'symbolic' is a boolean, and whether the attribute names are consistent.
It is not checked whether attribute 'theta' is present, because this parameter may be NULL in which case it is
not listed.
}
\author{
Martijn van Noort
}
