% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{isValidVariations}
\alias{isValidVariations}
\title{Check validity of a variational matrix}
\usage{
isValidVariations(df, nmdf = "input 'df'")
}
\arguments{
\item{df}{Any object.}

\item{nmdf}{Name (string) by which this object should be referred in any error messages. By default "input 'df'".}
}
\value{
\code{NULL} if \code{df} is a valid data frame with a variational matrix,
and an error message or list of error messages if not.
}
\description{
Checks whether the provided input is a valid variational matrix,
that is, whether it is in the format as produced by \code{\link{calcVariations}}, \code{\link{calcFirstVariations}}
or \code{\link{calcVariationsFim}}.
In particular, it is checked whether the input is a data frame of numeric values, with attributes 'type', 'symbolic',
'secOrd', where 'type' is set to 'VariationalMatrix' and 'symbolic' and 'secOrd' are booleans, and with columns
't', 'i', 'y', 'dy_d<v1>' and (if second order) 'd2y_d<v1>_<v2>', where variables v1 and v2 are replaced by names
that need to be present in 'theta', 'eta' or 'eps'.
The consistency of attributes 'theta', 'eta', 'eps' and 'parameterNormalized' is checked.
It is not checked whether attributes 'theta', 'eta', 'eps', 'parameterNormalized' and 'outputNormalized' are present,
because these parameters may be NULL in which case they are not listed.
}
\author{
Martijn van Noort
}
