% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{normalizeVariations}
\alias{normalizeVariations}
\title{Normalize variations}
\usage{
normalizeVariations(df, parNorm = TRUE, outNorm = TRUE)
}
\arguments{
\item{df}{A data frame containing a variational matrix, optionally with second order derivatives,
as produced by \code{\link{calcVariations}}, \code{\link{calcFirstVariations}} or
\code{\link{calcVariationsFim}}, that is, with columns 't', 'i', 'y', 'dy_d<v1>' and
(if second order) 'd2y_d<v1>_<v2>', where variables v1 and v2 are replaced by names.
Optionally this matrix may be parameter- or output-normalized.}

\item{parNorm}{Vector (not necessarily named) of booleans of length equal to the number of parameters used to
construct \code{df}, i.e., those listed in its attributes 'theta', 'eta' and 'eps'.
The value \code{TRUE} means the corresponding individual parameter should be normalized in the FIM,
\code{FALSE} means not.
Instead of a vector, may also provide a single boolean, where \code{TRUE} (default) stands for a vector of
all \code{TRUE}, and \code{FALSE} for a vector of all \code{FALSE}.
May also provide a character vector with the names of the parameters to be normalized.
Invalid names are ignored, with a warning.}

\item{outNorm}{\code{TRUE} (default) if variations should be output-normalized, \code{FALSE} if not.}
}
\value{
A data frame in same format, where the columns 'dy_d<v1>' contain the values, normalized as specified.
The function displays an error and returns \code{NULL} if the input is not in the specified format,
\code{parNorm} does not contain information on all parameters, or \code{outNorm = TRUE} and the 'y' column
of \code{df} contains zeroes.
The attributes are as those of \code{df}, with 'parameterNormalized' and 'outputNormalized' adapted according
to the new normalization.
}
\description{
Applies parameter- and/or output-normalization to a given first and (optionally) second order variational matrix.
For the first order matrix, parameter-normalization is performed by right multiplication with a matrix
P = diag(p), where p is a vector of parameter values (if the parameter is to be normalized) or 1 (if not),
and the output-normalization involves left multiplication with a matrix Y^(-1) where Y = diag(y) and y is
the vector of output values.
For the second order matrix (actually tensor), parameter-normalization involves two suitable multiplications
with P, and the output-normalization is done in the same way as for the first order matrix.
If the matrix was already normalized one way or the other, then the normalization is adapted to the new settings.
Parameter-normalization is selected per parameter, while output-normalization is a single choice.
}
\author{
Martijn van Noort
}
