% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parid-package.R
\docType{package}
\name{parid-package}
\alias{parid}
\alias{parid-package}
\title{parid: Calculation of parameter identifiability indicators for nonlinear mixed effects models}
\description{
Methods for determining the identifiability of nonlinear mixed effects model parameters, namely: Sensitivity Matrix Method (SMM), Fisher Information Matrix Method (FIMM) and Aliasing. The model is given as a set of differential equations, with specification of observed outputs, parameters and dosing. The methods assess whether the parameters can be identified from the model output, and optionally indicate dependencies between parameters.
}
\author{
\strong{Maintainer}: Martijn van Noort \email{info@lapp.nl}

Other contributors:
\itemize{
  \item LAP&P Consultants BV [copyright holder]
}

}
\keyword{internal}
