% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aliasing.r
\name{plotAliasing}
\alias{plotAliasing}
\title{Plot of normalized sensitivities and aliasing scores}
\usage{
plotAliasing(aliasing, elt = NULL, vars = NULL)
}
\arguments{
\item{aliasing}{Named list of aliasing score output as produced by \code{\link{calcAliasingScoresFromMatrix}}
or \code{\link{calcAliasingScoresFromModel}}.}

\item{elt}{List with elements "S", "P", "A" and/or "T" indicating which plot(s) should be made.
\code{NULL} can be used as shorthand for all.}

\item{vars}{Vector of names of parameters to be included in the plots.
If \code{NULL} (default) or missing, all parameters in aliasing are used.}
}
\value{
A named list of plots with the same names as in elt.
Element "S" is a time evolution plot of the normalized sensitivities.
Element "P" is a time evolution plot of the absolute value of the normalized sensitivities.
Element "A" is a heatmap of the aliasing scores.
Element "T" is like "A", but also showing the aliasing scores as text.
}
\description{
Generates several plots for normalized sensitivities and/or aliasing scores, as computed by
\code{\link{calcAliasingScoresFromMatrix}} or \code{\link{calcAliasingScoresFromModel}}.
}
\author{
Martijn van Noort
}
