% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{plotFim}
\alias{plotFim}
\title{Plot the FIM as a heatmap}
\usage{
plotFim(fim, vars = NULL, log = TRUE, label = FALSE)
}
\arguments{
\item{fim}{A matrix with row and column names containing the FIM,
as produced by \code{\link{calcFimFromMatrix}} or \code{\link{calcFimFromModel}}.}

\item{vars}{Vector of names of parameters to be included in the plots.
If \code{NULL} (default) or missing, all parameters in the FIM are used.}

\item{log}{If \code{TRUE} (default), then the FIM values are shown on the log scale.
If \code{FALSE}, then not.}

\item{label}{If \code{TRUE}, then the heatmap is labelled with the FIM values.
If \code{FALSE} (default), then not.}
}
\value{
A plot.
}
\description{
The given FIM is plotted as a heatmap, with one tile for each matrix element.
}
\details{
If on log scale, the heatmap uses separate color gradients for positive and negative values, if possible.
If this is not possible (it requires the package \code{ggnewscale}),
then negative values are indicated by a '-' symbol.
Optionally, the heatmap is labelled with the matrix values.
}
\seealso{
Other plotting and printing: 
\code{\link{plotAliasing}()},
\code{\link{plotSensitivities}()},
\code{\link{plotVariations}()},
\code{\link{print.keepattr}()}
}
\author{
Martijn van Noort
}
\concept{plotting and printing}
