% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.r
\name{plotSensitivities}
\alias{plotSensitivities}
\title{Plot sensitivity results}
\usage{
plotSensitivities(sens, elt = NULL)
}
\arguments{
\item{sens}{Named list of sensitivity indicators as produced by \code{\link{calcSensitivityFromMatrix}},
\code{\link{calcSensitivityFromModel}} or \code{\link{simplifySensitivities}}.}

\item{elt}{List with elements "R", "M" and/or "L" indicating which plot(s) should be made.
\code{NULL} can be used as shorthand for all.}
}
\value{
A named list of plots with one element for each element in elt that is present in sens.
In case of error, the return value is \code{NULL}.
}
\description{
Generates plots for one or more of the "R", "M" and "L" elements of the output of
\code{\link{calcSensitivityFromMatrix}}, \code{\link{calcSensitivityFromModel}} or \code{\link{simplifySensitivities}}.
For "R" or "M", the plot shows the parameter relations, while for "L" it shows the ordering of the parameters.
}
\seealso{
Other plotting and printing: 
\code{\link{plotAliasing}()},
\code{\link{plotFim}()},
\code{\link{plotVariations}()},
\code{\link{print.keepattr}()}
}
\author{
Martijn van Noort
}
\concept{plotting and printing}
