% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{plotVariations}
\alias{plotVariations}
\title{Plot variations}
\usage{
plotVariations(
  df,
  outNames = NULL,
  plotModel = TRUE,
  plotFirst = TRUE,
  plotSecond = TRUE,
  plotIIV = FALSE,
  plotRes = TRUE,
  plotPoints = FALSE
)
}
\arguments{
\item{df}{A data frame containing a variational matrix with or without second order derivatives,
as produced by \code{\link{calcVariations}}, \code{\link{calcFirstVariations}} or \code{\link{calcVariationsFim}},
that is, with columns 't', 'i', 'y', 'dy_d<v1>' and optionally 'd2y_d<v1>_<v2>', where variables v1 and v2
are replaced by names.}

\item{outNames}{A vector of strings containing the names of the outputs y_i in order.
They are used as labels.
If \code{NULL} (default), then the index \code{i} is used as label.}

\item{plotModel}{If \code{TRUE} (default), the model is plotted, if \code{FALSE} then not.}

\item{plotFirst}{If \code{TRUE} (default), the first variations are plotted, if \code{FALSE} then not.}

\item{plotSecond}{If \code{TRUE} (default), the second variations are plotted if present, if \code{FALSE} then not.}

\item{plotIIV}{If \code{TRUE}, the variations with respect to IIV variables are plotted if present, if \code{FALSE} (default) then not.
This requires the attribute 'eta' to be present.}

\item{plotRes}{If \code{TRUE} (default), the variations with respect to residual variables are plotted if present, if \code{FALSE} then not.
This requires the attribute 'eps' to be present.}

\item{plotPoints}{If \code{FALSE} (default), only lines are plotted, if \code{TRUE} also plots points.}
}
\value{
A plot.
In case of error, an error message is printed and the return value is \code{NULL}.
}
\description{
Creates plots of the model and its variations, for a given variational matrix.
}
\seealso{
Other plotting and printing: 
\code{\link{plotAliasing}()},
\code{\link{plotFim}()},
\code{\link{plotSensitivities}()},
\code{\link{print.keepattr}()}
}
\author{
Martijn van Noort
}
\concept{plotting and printing}
