% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.r
\name{print.keepattr}
\alias{print.keepattr}
\title{Print values}
\usage{
\method{print}{keepattr}(x, ...)
}
\arguments{
\item{x}{object to be printed.}

\item{...}{other arguments passed to the print method.}
}
\value{
\code{x}
}
\description{
This is a special version for objects (in particular data frames or matrices) with the 'keepattr' class.
It removes the 'keepattr' class and prints according to the remaining class specification.
If \code{x} is a variational matrix or FIM, it is printed without attributes.
}
\seealso{
Other plotting and printing: 
\code{\link{plotAliasing}()},
\code{\link{plotFim}()},
\code{\link{plotSensitivities}()},
\code{\link{plotVariations}()}
}
\author{
Martijn van Noort
}
\concept{plotting and printing}
