% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.r
\name{processErrors}
\alias{processErrors}
\title{Process errors}
\usage{
processErrors(errors)
}
\arguments{
\item{errors}{A vector of errors, of the form <caller>: <message>}
}
\value{
\code{NULL}
}
\description{
Process a vector of errors, by printing them to stdout.
}
\author{
Martijn van Noort
}
