% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.r
\name{processWarns}
\alias{processWarns}
\title{Process warnings}
\usage{
processWarns(warns)
}
\arguments{
\item{warns}{A vector of warnings, of the form <caller>: <message>}
}
\value{
\code{NULL}
}
\description{
Process a vector of warnings, by printing them to stdout.
}
\author{
Martijn van Noort
}
