% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{removeEmptyRowCols}
\alias{removeEmptyRowCols}
\title{Removes empty rows and columns from a symmetric matrix}
\usage{
removeEmptyRowCols(mat, varmat)
}
\arguments{
\item{mat}{The numeric matrix to check.}

\item{varmat}{The boolean matrix to check.}
}
\value{
A list of two elements, namely the adapted versions of \code{mat} and \code{varmat}.
}
\description{
Removes rows and columns from the input matrices \code{mat} and \code{varmat} if they contain only
zeroes in \code{mat} and only \code{FALSE} in \code{varmat}.
The input matrics are assumed to be symmetric.
This can be used to reduce variance-covariance matrices.
}
\author{
Martijn van Noort
}
