% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{simplifyFimIdent}
\alias{simplifyFimIdent}
\title{Simplify FIM identifiability results}
\usage{
simplifyFimIdent(fimId, tol = 0.001)
}
\arguments{
\item{fimId}{Named list of FIM identifiability indicators as produced by \code{\link{fimIdent}}.}

\item{tol}{Threshold for setting elements to 0.
Default value 0.001, i.e. all elements contributing less than 0.1\% to the Euclidean vector norm are set to 0.}
}
\value{
The list \code{fimId}, with a modified "directions" element.
If the input list \code{fimId} does not contain this element, then it is returned without change.
}
\description{
Simplify the "directions" element of the output of \code{\link{fimIdent}}, by setting all elements
smaller than a threshold (in absolute value) to 0.
The resulting vectors are scaled to their original norms.
}
\seealso{
Other result modifiers: 
\code{\link{normalizeFim}()},
\code{\link{normalizeVariations}()},
\code{\link{simplifySensitivities}()}
}
\author{
Martijn van Noort
}
\concept{result modifiers}
