% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.r
\name{simplifySensitivities}
\alias{simplifySensitivities}
\title{Simplify sensitivity results}
\usage{
simplifySensitivities(sens, elt = NULL, tol = 0.001)
}
\arguments{
\item{sens}{Named list of sensitivity indicators as produced by \code{\link{calcSensitivityFromMatrix}} or
\code{\link{calcSensitivityFromModel}}.}

\item{elt}{List with elements "R", "M" and/or "L" for simplification of R, M and/or L.
\code{NULL} can be used as shorthand for all.}

\item{tol}{Threshold for setting elements to 0.
Default value 0.001, i.e. all elements contributing less than 0.1\% to the Euclidean vector norm are set to 0.}
}
\value{
The list \code{sens}, with a modified "R", "M" and/or "L" element.
If the input list \code{sens} does not contain these elements, then it is returned without change.
}
\description{
Simplify the "R", "M" and/or "L" element of the output of \code{\link{calcSensitivityFromMatrix}} or
\code{\link{calcSensitivityFromModel}}, by setting all elements smaller than a threshold (in absolute value) to 0.
The results are renormalized.
}
\author{
Martijn van Noort
}
