% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.r
\name{[.keepattr}
\alias{[.keepattr}
\title{Extract or replace parts of an object, keeping attributes}
\usage{
\method{[}{keepattr}(x, ...)
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s).}

\item{...}{other arguments passed to a parent generic. Typically used to specify a subset.}
}
\value{
The specified subset of \code{x}. If \code{x} is a variational matrix or FIM, and the result
of subsetting is not a valid variational matrix or FIM, then the attributes will be lost.
This can happen for example when the FIM is no longer a square matrix.
They will also be lost if the class changes.
This can happen for example when a single element is extracted.
}
\description{
This is a special version for objects (in particular data frames or matrices) with the 'keepattr' class,
that will keep attributes on extraction.
If the object \code{x} is a variational matrix or FIM, they will be kept only if the result is still a valid
variational matrix or FIM, respectively.
}
\seealso{
Other retrieval functions: 
\code{\link{getAllParsFim}()},
\code{\link{getAllParsVariations}()},
\code{\link{getParVecFim}()},
\code{\link{getParVecVariations}()},
\code{\link{getParsFim}()},
\code{\link{getParsVariations}()},
\code{\link{subFim}()}
}
\author{
Martijn van Noort
}
\concept{retrieval functions}
