% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{subFim}
\alias{subFim}
\title{Subset a FIM matrix}
\usage{
subFim(fim, inds)
}
\arguments{
\item{fim}{A matrix with row and column names containing the FIM,
as produced by \code{\link{calcFimFromMatrix}} or \code{\link{calcFimFromModel}}.}

\item{inds}{Vector of indices of elements to keep. They may be provided as numerals or names.}
}
\value{
The submatrix \code{fim[inds, inds]}.
Returns \code{NULL} and prints an error if the input \code{fim} is not a valid FIM.
Returns the input \code{fim} and prints an error if the result of subsetting is not a valid FIM.
This should not be possible.
}
\description{
\code{subFim(fim, inds)} works just like \code{fim[inds, inds]}.
Use \code{subFim} instead of normal matrix subsetting functions to preserve attributes.
}
\author{
Martijn van Noort
}
