% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherInfo.r
\name{traceprod}
\alias{traceprod}
\title{Calculate the trace of the product of two given matrices}
\usage{
traceprod(mat1, mat2)
}
\arguments{
\item{mat1}{A numeric matrix.}

\item{mat2}{Another numeric matrix, of the same dimension.}
}
\value{
tr(mat1^t mat2).
}
\description{
The trace of the "skew" product of equal dimension matrices A and B is given by
tr(A^t B) = sum_ij A_ij B_ij.
}
\author{
Martijn van Noort
}
