% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationalEq.r
\name{transformCall}
\alias{transformCall}
\title{Transform a call by rpelacing variables by functions or vice versa}
\usage{
transformCall(callExpr, pList)
}
\arguments{
\item{callExpr}{The call to transform.}

\item{pList}{List of parameters to be used as function arguments. Each element is a language object.}
}
\value{
The transformed call.
}
\description{
Transforms a call (e.g. a function body) by replacing state variables by state functions or vice versa.
The state variables are assumed to be of the form 'y[]', where state is a unique identifier
of the component, of the form '<Component>\emph{<Deriv1>}<Deriv2>', where the component may be a name or a number,
and the derivatives are optional.
}
\author{
Martijn van Noort
}
